Applied changes requested
