+++ kura/examples/org.eclipse.kura.example.wire.math.trig/build.properties
