--- kura/examples/org.eclipse.kura.example.wire.trigonometric.functions.provider/OSGI-INF/TrigonometricComponents.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Copyright (c) 2020 Eurotech and/or its affiliates and others All rights 
-	reserved. This program and the accompanying materials are made available 
-	under the terms of the Eclipse Public License v1.0 which accompanies this 
-	distribution, and is available at http://www.eclipse.org/legal/epl-v10.html -->
-<scr:component
-	xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0"
-	name="org.eclipse.kura.wire.TrigonometricOperators" activate="activate"
-	deactivate="deactivate" modified="updated" enabled="true"
-	immediate="true" configuration-policy="require">
-	<implementation
-		class="org.eclipse.kura.example.wire.trigonometric.functions.provider.TrigonometricComponent" />
-	<service>
-		<provide
-			interface="org.eclipse.kura.configuration.ConfigurableComponent" />
-		<provide interface="org.eclipse.kura.wire.WireComponent" />
-		<provide interface="org.osgi.service.wireadmin.Consumer" />
-		<provide interface="org.eclipse.kura.wire.WireEmitter" />
-		<provide interface="org.eclipse.kura.wire.WireReceiver" />
-		<provide interface="org.osgi.service.wireadmin.Producer" />
-	</service>
-	<reference bind="bindWireHelperService" cardinality="1..1"
-		interface="org.eclipse.kura.wire.WireHelperService"
-		name="WireHelperService" policy="static"
-		unbind="unbindWireHelperService" />
-	<property name="service.pid"
-		value="org.eclipse.kura.wire.Average" />
-	<property name="kura.ui.service.hide" type="Boolean"
-		value="true" />
-	<property name="input.cardinality.minimum" type="Integer"
-		value="1" />
-	<property name="input.cardinality.maximum" type="Integer"
-		value="1" />
-	<property name="input.cardinality.default" type="Integer"
-		value="1" />
-	<property name="output.cardinality.minimum" type="Integer"
-		value="1" />
-	<property name="output.cardinality.maximum" type="Integer"
-		value="1" />
-	<property name="output.cardinality.default" type="Integer"
-		value="1" />
-</scr:component>
