+++ kura/examples/test/org.eclipse.kura.example.wire.math.trig.test/build.properties
