--- kura/examples/test/org.eclipse.kura.example.wire.trigonometric.functions.provider.test/OSGI-INF/TrigonometricComponentTest.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2020 Eurotech and others
-     All rights reserved. This program and the accompanying materials
-     are made available under the terms of the Eclipse Public License v1.0
-     which accompanies this distribution, and is available at
-     http://www.eclipse.org/legal/epl-v10.html
--->
-
-<scr:component
-	xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0"
-	activate="activate"
-	configuration-policy="ignore"
-	enabled="true"
-	immediate="true"
-	name="org.eclipse.kura.example.wire.trigonometric.functions.provider.test.TrigonometricComponentTest">
-	<implementation
-		class="org.eclipse.kura.example.wire.trigonometric.functions.provider.test.TrigonometricComponentTest" />
-	<property
-		name="service.pid"
-		value="org.eclipse.kura.example.wire.trigonometric.functions.provider.test.TrigonometricComponentTest" />
-	<reference
-		bind="bindWireGraphService"
-		cardinality="1..1"
-		interface="org.eclipse.kura.wire.graph.WireGraphService"
-		name="WireGraphService"
-		policy="static" />
- <reference bind="bindConfigurationService" cardinality="1..1" interface="org.eclipse.kura.configuration.ConfigurationService" name="ConfigurationService" policy="static"/>
-</scr:component>
