+++ kura/examples/test/pom.xml
@@ -276,6 +276,7 @@
         <module>org.eclipse.kura.example.testutil</module>
         <module>org.eclipse.kura.example.wire.logic.multiport.provider.test</module>
         <module>org.eclipse.kura.example.wire.trigonometric.functions.provider.test</module>
+        <module>org.eclipse.kura.example.wire.math.trig.test</module>
     </modules>
 
     <repositories>
